episode id as decimal + delete toast

+++ AnimApp/app/src/main/java/com/jony635/animapp/Episode.java
@@ -2,6 +2,6 @@ package com.jony635.animapp;
 
 public class Episode
 {
-    public int id = 0;
+    public float id = 0;
     public String link = "";
 }
